+++ src/vi/chat/chatentrywidget.py
