+++ src/vi/map/MapViewPage.py
