+++ src/vi/map/PanningWebView.py
@@ -1,5 +1,20 @@
-#   Vintel - Visual Intel Chat Analyzer
-#   Copyright (c) 2019. Steven Tschache (github@tschache.com)
+#     Vintel - Visual Intel Chat Analyzer
+#     Copyright (c) 2019. Steven Tschache (github@tschache.com)
+#
+#     This program is free software: you can redistribute it and/or modify
+#     it under the terms of the GNU General Public License as published by
+#     the Free Software Foundation, either version 3 of the License, or
+#     (at your option) any later version.
+#
+#     This program is distributed in the hope that it will be useful,
+#     but WITHOUT ANY WARRANTY; without even the implied warranty of
+#     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.	 See the
+#     GNU General Public License for more details.
+#
+#     You should have received a copy of the GNU General Public License
+#     along with this program.	 If not, see <http://www.gnu.org/licenses/>.
+#
+#
 #
 #   This program is free software: you can redistribute it and/or modify
 #   it under the terms of the GNU General Public License as published by
@@ -22,7 +37,7 @@ from PyQt5.QtCore import pyqtSignal, QUrl, QEvent, Qt
 from PyQt5.QtWidgets import QApplication
 from PyQt5 import QtGui
 from PyQt5.QtWebEngineWidgets import QWebEngineView, QWebEnginePage
-from vi.MapViewPage import MapViewPage
+from vi.map.MapViewPage import MapViewPage
 import logging
 
 LOGGER = logging.getLogger(__name__)
