+++ src/vi/ship/Ship.py
@@ -19,7 +19,7 @@
 
 from PyQt5.QtWidgets import QWidget
 import vi.ship.ShipDialog
-from vi.esihelper import EsiHelper
+from vi.esi.esihelper import EsiHelper
 import logging
 
 # TODO: show some information about the Ship...
