--- src/vi/test.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-
-</head>
-<body>
-</body>
-
-</html>
