--- src/vi/test1.html
@@ -1,140 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<svg height="1100" id="svgdoc" onload="init(evt)" version="1.1" viewbox="-550 0 1650 1100" width="2200" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
-<defs id="defs2974"></defs>
-<style id="style3" type="text/css"><![CDATA[
-/* jumps */
-.j	{ stroke: #000000; stroke-width: 1; }
-.jc	{ stroke: #FF0000; stroke-width: 1; }
-.jr	{ stroke: #800080; stroke-width: 1; }
-.jx	{ stroke: #000000; stroke-width: 3; }
-.jxc	{ stroke: #FF0000; stroke-width: 3; }
-.jxr	{ stroke: #800080; stroke-width: 3; }
-.jb	{ stroke: #0000FF; stroke-width: 3; fill: none; }
-
-/* external system */
-.eh	{ stroke: #00B000; stroke-width: 1; fill: #CCCCCC; }
-.el	{ stroke: #F00000; stroke-width: 1; stroke-dasharray: 8, 4; fill: #CCCCCC; }
-.e	{ stroke: #000000; stroke-width: 1; fill: #CCCCCC; }
-.exh	{ stroke: #00B000; stroke-width: 2; fill: #CCCCCC; }
-.exl	{ stroke: #F00000; stroke-width: 2; stroke-dasharray: 8, 3; fill: #CCCCCC; }
-.ex	{ stroke: #000000; stroke-width: 2; fill: #CCCCCC; }
-
-/* external system text */
-.es	{ font-size: 9px; font-family: Arial, Helvetica, sans-serif; fill: #000000; }
-.er	{ font-weight: bold; font-size: 7px; font-family: verdana, Arial, sans-serif; fill: #000000; }
-
-/* system */
-.sh	{ stroke: #00B000; stroke-width: 1; }
-.sl	{ stroke: #F00000; stroke-width: 1; stroke-dasharray: 8, 3; }
-.s	{ stroke: #000000; stroke-width: 1; }
-.sc	{ stroke-width: 1.2; stroke-dasharray: 4, 3; }
-.sx	{ stroke-width: 2.5; }
-
-.scl	{ stroke: transparent; stroke-width: 2; fill: transparent; }
-
-/* system text */
-.ss	{ font-family: Arial, Helvetica, sans-serif; font-size: 9px; fill: #000000; }
-.st	{ font-weight: bold; font-size: 7px; font-family: verdana, Arial, sans-serif; fill: #000000; }
-.so	{ fill: #ff0000; }
-.i	{ stroke: #0080FF; stroke-width: 1; fill: none; }
-
-/* station */
-.o	{ fill: #ffffff; }
-.oc	{ fill: #0000FF; }
-.oa	{ fill: #FFDC00; }
-.og	{ fill: #00FF00; }
-.om	{ fill: #FF0000; }
-.o1	{ fill: #ffffff; stroke-width: 0.5; stroke: #000000; }
-.o2	{ fill: #ffffff; stroke-width: 0.5; stroke: #000000; }
-.oo	{ fill: none; stroke: #000000; stroke-width: 1; }
-
-/* station services */
-.v1	{ fill: #FF0000; }
-.v2	{ fill: #00FF00; }
-.v3	{ fill: #FFDC00; }
-.v4	{ fill: #0000FF; }
-
-/* legend */
-.lb	{ stroke-width: 1; stroke: #000000; fill: #ffffff; }
-.lbt	{ stroke-width: 1; stroke: #000000; fill: #ffffff; }
-.lt	{ fill: #000000; font-family: Arial, Helvetica, sans-serif; font-weight: bold; font-size: 12px; }
-.lc	{ fill: #000000; font-family: Arial, Helvetica, sans-serif; font-size: 11px; }
-.l	{ fill: #000000; font-family: Arial, Helvetica, sans-serif; font-size: 10px; }
-
-/* highlight */
-.h	{ fill: none; stroke-width: 3; stroke: #000000; }
-.hw	{ fill: none; stroke-width: 2; stroke: #00AA00; }
-.hl	{ fill: none; stroke-width: 2; stroke: #F00000; }
-.hg	{ fill: none; stroke-width: 2; stroke: #00B000; }
-
-/* incursion */
-.inc	{ fill: url(#incBg); }
-.incs	{ fill: url(#incStBg); }
-
-/* contested */
-.con	{ fill: url(#conBg); }
-
-/* sov campaign triggered */
-.cam		{ fill: url(#camBg); }
-.cam-active	{ fill: url(#camActiveBg); }
-]]></style>
-<script id="script2968" type="text/javascript" xlink:href="http://s.dotlan.net/lib/dotSvg.js?1440663978"></script>
-<script id="script2970" type="text/javascript"><![CDATA[
-
-function init(E) {
-	SVGMap.init(E);
-}
-
-]]></script>
-<style id="style5" type="text/css"><![CDATA[
-.stopwatch {
-fill: #ffd39f; 
-stroke-width: 3; 
-stroke: #000000;
-}
-.es	{ font-size: 9px; font-family: Arial, Helvetica, sans-serif; fill: #000000; }
-.er	{ font-weight: bold; font-size: 7px; font-family: verdana, Arial, sans-serif; fill: #000000; }
-]]></style><div><symbol dummy="_self" id="def1"><a id="a1" target="_self" xlink:href=""><rect class="stopwatch" height="60" id="rect1" rx="111" ry="111" width="100" x="100" y="100"></rect><text class="es" id="timer1" text-anchor="middle" x="114" y="110">Time 1</text><text class="er" style="fill: #000000" text-anchor="middle" x="128" y="127">?</text></a></symbol></div><script type="text/javascript">
-function startTimerCountdown(seconds, display) {
-    var start = new Date().getTime() + seconds * 1000, elapsed = '0.0';
-    window.setInterval(function() {
-        var time = start - new Date().getTime();
-        elapsed = Math.ceil(time / 100) / 10;
-        
-        if (elapsed < 0) {
-            return;
-        }
-        minutes = parseInt(elapsed / 60, 10);
-        seconds = parseInt(elapsed % 60, 10);
-
-        minutes = minutes < 10 ? "0" + minutes : minutes;
-        seconds = seconds < 10 ? "0" + seconds : seconds;
-
-        display.textContent = minutes + ":" + seconds;
-    }, 1000);
-}
-function startTimer(secondsMax, display) {
-    var end = new Date().getTime() + secondsMax * 1000, elapsed = '0.0', start = new Date().getTime();
-    window.setInterval(function() {
-        var time = new Date().getTime();
-        elapsed = Math.ceil((end - time) / 100) / 10;
-        
-        if (elapsed < 0) {
-            return;
-        }
-        elapsed = (time - start) / 1000;
-        minutes = parseInt(elapsed / 60, 10);
-        seconds = parseInt(elapsed % 60, 10);
-
-        minutes = minutes < 10 ? "0" + minutes : minutes;
-        seconds = seconds < 10 ? "0" + seconds : seconds;
-
-        display.textContent = minutes + ":" + seconds;
-    }, 1000);
-}
-window.onload = function () {
-var field1 = document.querySelector('#timer1');
-var delay1 = 1 * 60;
-startTimer(delay1, field1);
-
-}</script></svg>
